% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_utilities.R
\name{best_model}
\alias{best_model}
\title{Model Selection from ENMeval Output}
\usage{
best_model(sdm.outputs, data.dir)
}
\arguments{
\item{sdm.outputs}{List of ENMeval objects containing model results, as produced by `load_results()`}

\item{data.dir}{Character string specifying path to directory to save the model selection RDS output}
}
\value{
data.frame with parameters of selected optimal model
}
\description{
Selects optimal model from ENMeval outputs using sequential criteria
}
\details{
Uses sequential criteria from Kass 2021:
  1. Minimizes omission rate (or.10p.avg)
  2. Maximizes CBI validation score (cbi.val.avg)
  3. If multiple models tie, randomly select one of the models (statistically equivalent)
}
\examples{
\dontrun{
# sdm.outputs is a list of ENMeval objects, as produced by load_results()
modelSelection <- best_model(sdm.outputs, data.dir = "path/to/outputs/")
}
}
\references{
Kass et al. 2021 https://doi.org/10.1111/2041-210X.13628, https://jamiemkass.github.io/ENMeval/articles/ENMeval-2.0-vignette.html#select
}
