% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_modelGeneration.R
\name{check_envCorr}
\alias{check_envCorr}
\title{Write Environmental Correlation Plot to File}
\usage{
check_envCorr(output.dir, env.vars, occ.sf.pts)
}
\arguments{
\item{output.dir}{Character string specifying output directory path}

\item{env.vars}{SpatRaster object containing environmental predictor layers}

\item{occ.sf.pts}{sf points object containing species occurrence locations}
}
\value{
None (writes plot file to disk)
}
\description{
Creates and saves correlation plot showing relationships between environmental variables
}
\details{
Extracts environmental values at occurrence points and creates correlation plot showing
         Pearson correlations between all pairs of variables using ggcorrplot
}
\examples{
\dontrun{
check_envCorr(output.dir = "path/to/output/", env.vars = env.vars, occ.sf.pts = occ.points)
}
}
