% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_utilities.R
\name{convert_thresh_toSf}
\alias{convert_thresh_toSf}
\title{Convert Thresholded Rasters to SF Polygons}
\usage{
convert_thresh_toSf(thresh.dir, output.dir, output.name, simp.factor)
}
\arguments{
\item{thresh.dir}{Character string specifying directory containing thresholded rasters}

\item{output.dir}{Character string specifying directory for output files}

\item{output.name}{Character string specifying name for output SF file}

\item{simp.factor}{Numeric simplification factor passed to `sf::st_simplify()` when converting rasters to polygons}
}
\description{
Converts a directory of thresholded binary rasters into a single SF polygon object
}
\details{
Converts each binary raster in the input directory to polygons and combines them
into a single SF object. Only keeps polygons where raster value equals 1
(suitable habitat). Raster filenames are used as taxon names in output.
}
\examples{
\dontrun{
convert_thresh_toSf("data/thresholded/", "output/polygons/", "habitat_polygons.shp")
}
}
