% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_pipelines.R
\name{maxent_nullModel_pipe}
\alias{maxent_nullModel_pipe}
\title{Null Model Pipeline for Maxent Models}
\usage{
maxent_nullModel_pipe(
  null.iter = 1000,
  data.dir = "data/maxent/",
  output.dir = "outputs/maxent/",
  parallel = F,
  numCores = 4
)
}
\arguments{
\item{null.iter}{Numeric, number of null model iterations to run}

\item{data.dir}{Character, path to directory containing model outputs}

\item{output.dir}{Character, path to directory for null model outputs}

\item{parallel}{Logical, whether to enable parallel processing}

\item{numCores}{Numeric, number of cores to use for parallel processing}
}
\value{
None. Saves null model outputs to disk:
  - Null model RDS files
  - Evaluation histograms
  - Evaluation violin plots
}
\description{
Generates and evaluates null models for Maxent models created with maxent_pipe()
}
