% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_modelPrediction.R
\name{overlap_ensemble_dist.dep}
\alias{overlap_ensemble_dist.dep}
\title{Calculate Overlap Between Historic and Future Suitability (deprecated)}
\usage{
overlap_ensemble_dist.dep(
  hist.pred.rast.thresh,
  future.thresh.ensemble,
  species.idx,
  time.idx,
  scenario.idx
)
}
\arguments{
\item{hist.pred.rast.thresh}{List of historic thresholded suitability SpatRasters, indexed by species}

\item{future.thresh.ensemble}{Nested list of future thresholded ensemble SpatRasters, indexed by species, time, then scenario}

\item{species.idx}{Numeric index identifying the species within `hist.pred.rast.thresh` and `future.thresh.ensemble`}

\item{time.idx}{Numeric index identifying the time period within `future.thresh.ensemble`}

\item{scenario.idx}{Numeric index identifying the scenario within `future.thresh.ensemble`}
}
\description{
Deprecated. Combines a historic thresholded suitability raster with a future thresholded ensemble raster to visualise loss/retention of habitat.
}
