% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_modelPrediction.R
\name{predict_maxentJar}
\alias{predict_maxentJar}
\title{Generate MaxEnt Prediction Using Maxent.jar}
\usage{
predict_maxentJar(env.vars, SDM.obj, model, taxon.name)
}
\arguments{
\item{env.vars}{SpatRaster of environmental predictor layers to predict onto}

\item{SDM.obj}{ENMeval object containing model results}

\item{model}{Character string, name of the selected model within SDM.obj (as used to index `eval.models(SDM.obj)`)}

\item{taxon.name}{Character string, name applied to the output raster layer}
}
\value{
SpatRaster of cloglog suitability predictions, named according to taxon.name
}
\description{
Projects a fitted maxent.jar model onto a set of environmental raster layers
}
\details{
Extracts the selected model from SDM.obj and projects it onto env.vars using `dismo::predict()` with cloglog output format
}
\examples{
\dontrun{
predict_maxentJar(env.vars, SDM.obj, model = "fc.LQ_rm.1", taxon.name = "species_a")
}
}
