% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_utilities.R
\name{split_futureEnvs_sf}
\alias{split_futureEnvs_sf}
\title{Split Future Prediction Naming into Components}
\usage{
split_futureEnvs_sf(sf.future.thresh)
}
\arguments{
\item{sf.future.thresh}{sf polygon object with a taxonName column formatted as "taxonName_time.period_ssp_model"}
}
\value{
sf polygon object with taxonName split into separate taxonName, time.period, ssp, and model columns
}
\description{
Splits the taxonName column of an sf polygon object into separate taxonName, time.period, ssp, and model columns
}
\details{
Uses `tidyr::separate()` on the taxonName column, retaining the original column
}
\examples{
\dontrun{
split_futureEnvs_sf(sf.future.thresh)
}
}
