% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_utilities.R
\name{stack_futureEnvs}
\alias{stack_futureEnvs}
\title{Reorganise Future Predictions into a Nested List}
\usage{
stack_futureEnvs(pred.rast.future)
}
\arguments{
\item{pred.rast.future}{SpatRaster stack of future predictions, with layer names formatted as "species_time_ssp_model"}
}
\value{
Nested list of SpatRaster layers structured as species > time > ssp > model
}
\description{
Splits a named SpatRaster stack of future predictions into a nested list structured as species > time > ssp > model
}
\details{
Layer names are split on underscores into their four components; assumes a fixed species_time_ssp_model naming convention
}
\examples{
\dontrun{
stack_futureEnvs(pred.rast.future)
}
}
