% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_modelGeneration.R
\name{subset_EnvByStudyArea}
\alias{subset_EnvByStudyArea}
\title{Subset Environmental Data by Study Area}
\usage{
subset_EnvByStudyArea(env.vars, study.area)
}
\arguments{
\item{env.vars}{SpatRaster object containing environmental variables}

\item{study.area}{sf object defining the study area boundary}
}
\value{
SpatRaster cropped to study area extent
}
\description{
Crops environmental raster data to the extent of a defined study area
}
\details{
This function reduces processing time by cropping environmental data
to only the relevant study area
}
\examples{
\dontrun{
env_subset <- subset_EnvByStudyArea(env_vars, study_area)
}
}
