% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_modelValidation.R
\name{write_response_curves}
\alias{write_response_curves}
\title{Write Response Curves to File}
\usage{
write_response_curves(SDM.obj, model, output.dir, data.dir)
}
\arguments{
\item{SDM.obj}{ENMeval object containing model results}

\item{model}{Character string specifying the name of the best model settings within `SDM.obj`, as selected by `best_model()`}

\item{output.dir}{Character string specifying output directory path}

\item{data.dir}{Character string specifying directory to write the intermediate response curve CSV to}
}
\value{
None (writes plot file to disk)
}
\description{
Creates and saves response curves plot for the best performing model
}
\details{
Creates response curves showing predicted suitability across the range of each environmental variable, using the best model selected by best_model()
}
\examples{
\dontrun{
write_response_curves(SDM.obj, model = "fc.LQ_rm.1",
                       output.dir = "path/to/output/", data.dir = "path/to/data/")
}
}
