% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_utilities.R
\name{write_sdm_rast}
\alias{write_sdm_rast}
\title{Write Best Model Raster to File}
\usage{
write_sdm_rast(output.dir, SDM.obj, model)
}
\arguments{
\item{output.dir}{Character string specifying output directory path}

\item{SDM.obj}{ENMeval object containing model results}

\item{model}{Character string specifying the name of the best model settings within `SDM.obj`, as selected by `best_model()`}
}
\value{
None (writes raster file to disk)
}
\description{
Writes the prediction raster from the best performing model to file
}
\examples{
\dontrun{
write_sdm_rast("path/to/output.dir", SDM.obj, model = "fc.LQ_rm.1")
}
}
