% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_modelValidation.R
\name{write_var_contrib}
\alias{write_var_contrib}
\title{Write Variable Contribution Plot to File}
\usage{
write_var_contrib(SDM.obj, model, output.dir, data.dir)
}
\arguments{
\item{SDM.obj}{ENMeval object containing model results}

\item{model}{Character string specifying the name of the best model settings within `SDM.obj`, as selected by `best_model()`}

\item{output.dir}{Character string specifying output directory path}

\item{data.dir}{Character string specifying directory to write the intermediate variable importance CSV to}
}
\value{
None (writes plot file to disk)
}
\description{
Creates and saves variable contribution plot for the best performing model
}
\details{
Creates a bar plot showing the percent contribution of each environmental variable to the best model selected by best_model()
}
\examples{
\dontrun{
write_var_contrib(SDM.obj, model = "fc.LQ_rm.1",
                   output.dir = "path/to/output/", data.dir = "path/to/data/")
}
}
