% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_pipelines.R
\name{maxent_modelGeneration_pipe}
\alias{maxent_modelGeneration_pipe}
\title{ENMeval Pipeline Main Function}
\usage{
maxent_modelGeneration_pipe(
  occ.sf.pts,
  env.vars,
  bg.areas,
  study.area,
  data.dir,
  seed = 343,
  java.mem.gb = 16,
  bg.buffer.dist = NULL,
  bg.pts.num = 10000,
  subset.col = "scientificName",
  clampStudyArea = T,
  raster.preds = T,
  doClamp = T,
  categoricals = NULL,
  partitions = "block",
  algorithm = "maxent.jar",
  overlap = T,
  overlapStat = c("D", "I"),
  tune.args = list(fc = c("L", "LQ", "LQH", "LQHP", "LQHPT"), rm = c(0.5, 1, 2, 4, 5)),
  enable.parallel = F,
  numCores = 4
)
}
\arguments{
\item{occ.sf.pts}{sf Points object, occurrence records}

\item{env.vars}{SpatRaster, environmental variables}

\item{bg.areas}{sf Polygon, background area polygons, such as ibra bioregions etc}

\item{study.area}{sf Polygon, study area polygon, must contain all occurrence points.}

\item{data.dir}{Character, path to data directory; subfolders (biasLayer, thinnedRecords, bgAreas, bgPoints, modelOutput, modelOutputRast) are created beneath it for intermediate and final model outputs}

\item{seed}{Numeric, random seed for reproducibility}

\item{java.mem.gb}{Numeric, Java memory in GB}

\item{bg.buffer.dist}{Numeric, buffer distance for background points}

\item{bg.pts.num}{Numeric, number of background points}

\item{subset.col}{Character, column name for subsetting data}

\item{clampStudyArea}{Logical, restrict background areas to within area of interest to reduce computation time}

\item{raster.preds}{Logical, whether to generate and save raster predictions for each model (default TRUE). If FALSE, the modelOutputRast subfolder is not created}

\item{doClamp}{Logical, whether to clamp predictions to avoid extrapolation beyond training data}

\item{categoricals}{Character vector, use to tell MaxEnt which of your variables are categorical. Used terra::rasterise and assign field for factor levels. Do not aggregate post as this will break layer for enmeval.}

\item{partitions}{Character, type of spatial partitioning}

\item{algorithm}{Character, modeling algorithm}

\item{overlap}{Logical, whether to calculate niche overlap}

\item{overlapStat}{Character, statistic for niche overlap}

\item{tune.args}{List, model tuning arguments}

\item{enable.parallel}{Logical, enable parallel processing}

\item{numCores}{Numeric, number of cores for parallel processing}
}
\value{
None. Saves model outputs to disk
}
\description{
Pipeline for standard workflow of generating maxent models from occurrence points, environmental raster stack, background areas and overall study area.
}
