% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxEnt_pipelines.R
\name{maxent_modelValidation_pipe}
\alias{maxent_modelValidation_pipe}
\title{Validation Pipeline for Maxent Models}
\usage{
maxent_modelValidation_pipe(
  env.vars,
  occ.sf.pts,
  functions.dir = maxent.tools,
  data.dir = "data/maxent/",
  output.dir = "outputs/maxent/",
  raster.preds = T,
  thresh.method = "maxsss"
)
}
\arguments{
\item{env.vars}{SpatRaster, environmental variables used in original models}

\item{functions.dir}{Character, path to directory containing validation functions}

\item{data.dir}{Character, path to directory containing model outputs}

\item{output.dir}{Character, path to directory for validation outputs}
}
\value{
None. Saves validation outputs to disk:
  - Environmental correlations
  - Prediction rasters
  - Variable contribution plots
  - Response curves
  - Thresholded predictions
  - Stacked prediction rasters
}
\description{
Runs validation analyses and generates output files for Maxent models created with maxent_pipe()
}
